added license expired and pod running status test cases
