+++ consul-operator/componenttest/crds/ ops.dac.nokia.com_licenceexpireds.yaml
@@ -1,0 +1,52 @@
+
+---
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  annotations:
+    controller-gen.kubebuilder.io/version: v0.6.1
+  creationTimestamp: null
+  name: licenceexpireds.ops.dac.nokia.com
+spec:
+  group: ops.dac.nokia.com
+  names:
+    kind: LicenceExpired
+    listKind: LicenceExpiredList
+    plural: licenceexpireds
+    singular: licenceexpired
+  scope: Namespaced
+  versions:
+    - name: v1alpha1
+      schema:
+        openAPIV3Schema:
+          description: LicenceExpired is the Schema for the licenceexpireds API
+          properties:
+            apiVersion:
+              description: 'APIVersion defines the versioned schema of this representation
+              of an object. Servers should convert recognized schemas to the latest
+              internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
+              type: string
+            kind:
+              description: 'Kind is a string value representing the REST resource this
+              object represents. Servers may infer this from the endpoint the client
+              submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
+              type: string
+            metadata:
+              type: object
+            spec:
+              description: LicenceExpiredSpec defines the desired state of LicenceExpired
+              type: object
+            status:
+              description: LicenceExpiredStatus defines the observed state of LicenceExpired
+              type: object
+          type: object
+      served: true
+      storage: true
+      subresources:
+        status: {}
+status:
+  acceptedNames:
+    kind: ""
+    plural: ""
+  conditions: []
+  storedVersions: []
